requirements.txt
streamlit
streamlit-authenticator
PyYAML
python-nmap
boto3
requests